backend/requirements.txt
akshare==1.16.76
fastapi==0.115.12
loguru==0.7.3
pandas==2.2.3
passlib==1.7.4
pydantic[email]==2.11.4
pydantic_settings==2.9.1
pytest==8.3.5
python-dotenv==1.1.0
python_jose==3.4.0
Requests==2.32.3
SQLAlchemy==2.0.40
tenacity==9.1.2
uvicorn==0.34.0
pymysql==1.1.1
aiomysql==0.2.0
redis==6.0.0
apscheduler==3.11.0
cryptography==44.0.3